Reorganized Tests
Dissolved the `src/tests.rs` file and moved unit tests into the module files they test, as suggested in [The Rust Book - Test Organization](https://doc.rust-lang.org/book/ch11-03-test-organization.html).
The two tests which access the cache 1000 times are considered integration tests and were moved to `tests/many-accesses.rs`.
This resolves #18
